Datenübertragung durch Shader gefixt (ziemlich blöd, aber geht)

+++ SimulateTheWorld.Graphics.Data/OpenGL/DataVertex.cs
@@ -5,10 +5,10 @@ namespace SimulateTheWorld.Graphics.Data.OpenGL;
 public struct DataVertex
 {
     public Vector3 position;
-    public int tileType;
-    public int terrainType;
+    public float tileType;
+    public float terrainType;
 
-    public DataVertex(Vector3 position, int tileType, int terrainType)
+    public DataVertex(Vector3 position, float tileType, float terrainType)
     {
         this.position = position;
         this.tileType = tileType;
