+++ SimulateTheWorld.Graphics.Resources/Rendering/Shaders/points.geom
@@ -5,19 +5,19 @@ layout(points, max_vertices = 1) out;
 
 in DATA
 {
-    int tileType;
-    int terrainType;
+    float tileType;
+    float terrainType;
     mat4 projection;
 } data_in[];
 
-flat out int tileType;
-flat out int terrainType;
+flat out float tileType;
+flat out float terrainType;
 
 void main()
 {
     tileType = data_in[0].tileType;
     terrainType = data_in[0].terrainType;
-    
+                
     gl_Position = gl_in[0].gl_Position * data_in[0].projection;
     EmitVertex();
